feat: Update fileStorage.js to add file selection functionality
This commit modifies the fileStorage.js file to add the ability to select a file for opening. The selectFile() method is implemented, which opens a dialog box for the user to choose a file to open. The selected file path is then returned. This functionality will be useful for retrieving data from a specific file.